fix(hail): MSVC narrowing cast on chain_log_emit payload size
`38 + motd_len` is computed as `int + size_t`, then implicitly
narrowed to the `uint16_t` parameter on chain_log_emit. MSVC strict
mode rejects the implicit narrowing under -Werror. Cast explicitly;
the value is bounded to 38 + 256 = 294 by construction.

+++ server/game_sim.c
@@ -4958,7 +4958,7 @@ void world_reset(world_t *w) {
         if (motd_len > 0) {
             memcpy(payload + 38, st->hail_message, motd_len);
         }
-        (void)chain_log_emit(w, st, CHAIN_EVT_OPERATOR_POST, payload, 38 + motd_len);
+        (void)chain_log_emit(w, st, CHAIN_EVT_OPERATOR_POST, payload, (uint16_t)(38 + motd_len));
     }
 
     rebuild_signal_chain(w);
